db/osm/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
shapely>=2.0.0
geojson>=3.1.0
python-dotenv>=1.0.0
